requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
colorlog==6.9.0
confection==0.1.5
conllu==6.0.0
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
datasets==4.1.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastuuid==0.12.0
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.35.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
leveldb==0.201
Levenshtein==0.27.1
litellm==1.77.1
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==2.5.0
networkx==3.5
nltk==3.9.1
nmslib-metabrainz==2.1.3
numpy==2.3.4
openai==1.108.0
openpyxl==3.1.5
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.3.0
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.1
pyarrow==21.0.0
pybind11==3.0.1
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
quickumls==1.4.2
quickumls_simstring==1.1.5.post1
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
scispacy==0.6.2
sentence-transformers==5.1.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
typer==0.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
unqlite==0.9.9
urllib3==2.5.0
wasabi==0.10.1
weasel==0.4.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0